Refactor: Move ActivationCounterComponent and add onSpecial effect
